Cardio log done for date picking.  Still need better support for pre API 24 users.  Fixed up the route field.

+++ app/src/main/java/com/nicholaslocicero/guiles/guilesfitnesstracker/model/Entities/CardioWorkout.java
@@ -34,11 +34,11 @@ public class CardioWorkout {
         this.route = route;
     }
 
-    public Integer getMinutes() {
+    public Float getMinutes() {
         return minutes;
     }
 
-    public void setMinutes(Integer minutes) {
+    public void setMinutes(Float minutes) {
         this.minutes = minutes;
     }
